+++ Examples/Python/src/Geometry.cpp
@@ -55,6 +55,25 @@ struct GeometryIdentifierHookBinding : public Acts::GeometryIdentifierHook {
 namespace Acts::Python {
 void addGeometry(Context& ctx) {
   auto m = ctx.get("main");
+
+  {
+    py::class_<Acts::GeometryIdentifier>(m, "GeometryIdentifier")
+        .def(py::init<>())
+        .def(py::init<Acts::GeometryIdentifier::Value>())
+        .def("setVolume", &Acts::GeometryIdentifier::setVolume)
+        .def("setLayer", &Acts::GeometryIdentifier::setLayer)
+        .def("setBoundary", &Acts::GeometryIdentifier::setBoundary)
+        .def("setApproach", &Acts::GeometryIdentifier::setApproach)
+        .def("setSensitive", &Acts::GeometryIdentifier::setSensitive)
+        .def("setExtra", &Acts::GeometryIdentifier::setExtra)
+        .def("volume", &Acts::GeometryIdentifier::volume)
+        .def("layer", &Acts::GeometryIdentifier::layer)
+        .def("boundary", &Acts::GeometryIdentifier::boundary)
+        .def("approach", &Acts::GeometryIdentifier::approach)
+        .def("sensitive", &Acts::GeometryIdentifier::sensitive)
+        .def("extra", &Acts::GeometryIdentifier::extra);
+  }
+
   {
     py::class_<Acts::Surface, std::shared_ptr<Acts::Surface>>(m, "Surface")
         .def("geometryId",
